move load script
