--- docs/src/modules/utils/loadScript.js
@@ -1,8 +1,0 @@
-export default function loadScript(src, position) {
-  const script = document.createElement('script');
-  script.setAttribute('async', '');
-  script.src = src;
-  position.appendChild(script);
-
-  return script;
-}
